requirements.txt
asgiref==3.8.1
certifi==2024.2.2
chardet==3.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-import-export==3.3.7
djangorestframework==3.15.1
et-xmlfile==1.1.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pytz==2024.1
PyYAML==6.0.1
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.4.4
tablib==3.5.0
xlrd==2.0.1
xlwt==1.3.0
